Move specialized C API handle definitions next to the high-level types.
